tca/requirements.txt
Django==1.6.5
argparse==1.2.1
djangorestframework==2.3.13
factory-boy==2.3.1
wsgiref==0.1.2
